+++ CHANGES.rst
@@ -9,6 +9,8 @@
 
 - Drop support for Python 2.7, 3.5, 3.6.
 
+- Drop deprecated support for ``python setup.py test``.
+
 
 4.2.1 (2022-08-25)
 ==================